move usecase models and dto's to separate directories
