+++ src/lib/core/port/primary/switch-account-input-port.ts
@@ -1,5 +1,5 @@
 import { IronSession } from "iron-session";
-import { SwitchAccountRequest } from "../../data/switch-account-usecase-models";
+import { SwitchAccountRequest } from "../../data/usecase-models/switch-account-usecase-models";
 
 /**
  * Provides an interface for the {@link SwitchAccountUseCase}.
