+++ src/lib/core/port/secondary/auth-server-gateway-output-port.ts
@@ -1,4 +1,4 @@
-import { UserPassLoginAuthServerDTO } from "../../data/auth-server-dto";
+import { UserPassLoginAuthServerDTO } from "../../data/dto/auth-server-dto";
 
 export default interface AuthServerGatewayOutputPort { 
     userpassLogin(username: string, password: string, account: string, vo: string): Promise<UserPassLoginAuthServerDTO>
