+++ src/pages/api/rseaccountusage.ts
@@ -3,7 +3,8 @@ import StreamGatewayOutputPort from "@/lib/core/port/secondary/stream-gateway-ou
 import appContainer from "@/lib/infrastructure/config/ioc/container-config";
 import GATEWAYS from "@/lib/infrastructure/config/ioc/ioc-symbols-gateway";
 import { PassThrough } from "node:stream";
-
+import { HTTPRequest } from "@/lib/common/stream/http";
+import { Response } from "node-fetch";
 
 export const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
 
@@ -16,7 +17,19 @@ const rseAccountUsageRoute = async (req: NextApiRequest, res: NextApiResponse) =
     res.setHeader('X-Accel-Buffering', 'no');
     res.setHeader('Connection', 'keep-alive');
     res.setHeader('Content-Encoding', 'none');
-    const responseStream: PassThrough | null = await streamingGateway.getJSONChunks(url)
+    const request: HTTPRequest = {
+        url: url,
+        method: 'GET',
+        headers: {
+            'Content-Type': 'application/x-json-stream',
+        },
+        body: null,
+    }
+    const responseStream: PassThrough | Response = await streamingGateway.getJSONChunks(request)
+    if(responseStream instanceof Response) {
+        res.status(responseStream.status).json(responseStream.statusText)
+        return
+    }
     responseStream?.on('data', async (chunk) => {
         res.write(chunk.toString() + '\n')
         await sleep(1000);
